Sample Project Clean Up Work
