--- examples/sample-hybrid-project/.avail/sample-hybrid/.gitignore
@@ -1,1 +1,0 @@
-/**/settings-local.json
