--- examples/sample-hybrid-project/.avail/sample-hybrid/artifact-plans.json
@@ -1,37 +1,0 @@
-[
-	{
-		"version": "2.0.0.alpha02",
-		"outputLocation": 
-		{
-			"locationType": "project",
-			"scheme": "JAR",
-			"path": "/Users/Rich/Development/avail-repos/avail/examples/sample-hybrid-project/build/libs/sample-hybrid-project-2.0.0.alpha02.jar"
-		},
-		"artifactType": "APPLICATION",
-		"implementationTitle": "Avail Sample Hybrid Application",
-		"jarMainClass": "org.availlang.sample.AppKt",
-		"artifactDescription": "This description goes into the Avail manifest in the jar!",
-		"digestAlgorithm": "SHA=256",
-		"rootNames": 
-		[
-			"avail-stdlib",
-			"other-root",
-			"my-avail-root"
-		],
-		"includedFiles": 
-		[
-		],
-		"jarFileLocations": 
-		[
-		],
-		"zipFileLocations": 
-		[
-		],
-		"directoryLocations": 
-		[
-		],
-		"customManifestItems": 
-		{
-		}
-	}
-]
