--- examples/sample-hybrid-project/.gitignore
@@ -1,11 +1,0 @@
-/**/.DS_Store
-/**/*.jpage
-/**/*.ws.kts
-/.gradle
-.swp
-build
-/out
-/debug
-hs_err_*.log
-**/*.repo
-*-local-state.json
