--- examples/sample-hybrid-project/.idea/.gitignore
@@ -1,20 +1,0 @@
-# Default ignored files
-/shelf/
-/workspace.xml
-# Editor-based HTTP Client requests
-/httpRequests/
-# Datasource local storage ignored files
-/dataSources/
-/dataSources.local.xml
-
-/jarRepositories.xml
-/kotlinc.xml
-/libraries
-/libraries-with-intellij-classes.xml
-/modules
-/modules.xml
-/ClojureProjectResolveSettings.xml
-/compiler.xml
-/GrepConsole.xml
-/misc.xml
-/**/*.iml
