--- examples/sample-hybrid-project/buildSrc/.gitignore
@@ -1,15 +1,0 @@
-/**/.DS_Store
-/**/*.jpage
-/**/*.ws.kts
-/.gradle
-.swp
-build
-/stacks
-/modules.dot
-/out
-/debug
-/documentation/docs/src_docs
-/documentation/docs/index.md
-/documentation/site
-hs_err_*.log
-/.idea
