+++ examples/sample-library-linking-project/.avail/.gitignore
@@ -1,0 +1,2 @@
+/**/*local-state*
+/**/*.backup
