+++ examples/sample-library-linking-project/.avail/sample-library-linking/avail-stdlib/settings-local.json
@@ -1,0 +1,6 @@
+{
+	"palette": "",
+	"loadModulesOnStartup": 
+	[
+	]
+}
