+++ examples/sample-library-linking-project/.avail/sample-library-linking/avail-stdlib/templates.json
