+++ examples/sample-library-linking-project/.avail/sample-library-linking/my-avail-root/templates.json
