+++ examples/sample-library-linking-project/.avail/sample-library-linking/other-root/settings-local.json
@@ -1,0 +1,6 @@
+{
+	"palette": "",
+	"loadModulesOnStartup": 
+	[
+	]
+}
