+++ examples/sample-library-linking-project/.avail/sample-library-linking/other-root/templates.json
