+++ examples/sample-library-linking-project/.avail/sample-library-linking/styles.json
