+++ examples/sample-library-linking-project/.avail/sample-library-linking/templates.json
