+++ examples/sample-library-linking-project/copyright.txt
