+++ examples/sample-library-linking-project/gradlew.bat
