+++ examples/sample-library-linking-project/readme/workbench.jpg
