+++ examples/sample-library-linking-project/roots/my-avail-root/App.avail/Foreign Interface.avail
