+++ examples/sample-library-linking-project/roots/my-avail-root/App.avail/Network.avail/Network.avail
