+++ examples/sample-library-linking-project/roots/my-avail-root/App.avail/avail-java-ffi-1.0.jar
