+++ examples/sample-library-linking-project/roots/other-root/Other.avail
