+++ examples/sample-library-linking-project/settings.gradle.kts
@@ -1,0 +1,2 @@
+rootProject.name = "sample-library-linking-project"
+
