+++ examples/sample-primitives-project/README.md
@@ -1,0 +1,19 @@
+SAMPLE PRIMITIVE PROJECT OVERVIEW
+--------------------------------------------------------------------------------
+
+![Version](https://img.shields.io/badge/v2.0.0.alpha01-0f824e)
+
+***NOTE: 2023.09.08 The current state of this example is: Expected to Work***
+
+*This uses alpha version tools that are in flux. This README will be updated*
+*to reflect accurate usage for the 2.0.0 production release of the Avail*
+*toolchain. Once the tools are finalized, this sample will be updated to a*
+*working state relative to the latest production release.*
+
+This represents a project that creates Avail primitives in a JVM language, 
+in this case, Kotlin, that is dynamically linked in Avail with Avail's 
+primitive linking system. This adds new primitives to Avail that can be used in
+Avail code. This project will be updated over time to reflect changes to 
+Avail as well as to further demonstrate how Avail can be used in a JVM project.
+
+_This is a work in progress!_
